requirements.txt
beautifulsoup4==4.14.3
google-genai==1.3.0
playwright==1.58.0
python-dotenv==1.0.1
requests==2.32.5
pytest==8.0.0
pydantic==2.10.6
